# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "movie-reviews-api",
        "description": "API documentation movie-reviews-api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/people": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "get people.",
                "operationId": "b2985cc1b5ff63acae04676e8de27da8",
                "parameters": [
                    {
                        "name": "search_term",
                        "in": "query",
                        "description": "search_term",
                        "schema": {
                            "format": "string",
                            "default": ""
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "number of page",
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "number of items in a page",
                        "schema": {
                            "format": "int64",
                            "default": 10
                        }
                    },
                    {
                        "name": "followers_count",
                        "in": "query",
                        "description": "filter followers_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "medias_count",
                        "in": "query",
                        "description": "filter medias_count",
                        "schema": {
                            "format": "int64",
                            "default": ""
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "filter gender",
                        "schema": {
                            "format": "string",
                            "default": "",
                            "enum": [
                                "male",
                                "female",
                                ""
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sort by",
                        "schema": {
                            "format": "string",
                            "default": "",
                            "enum": [
                                "newest-created",
                                "oldest-created",
                                "youngest",
                                "oldest",
                                "most-followers",
                                "least-followers",
                                "most-medias",
                                "least-medias"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "person"
                ],
                "summary": "create new person.",
                "operationId": "209c4a23e6aa244623d8a2a3a548f25b",
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "name of person",
                                        "default": ""
                                    },
                                    "url": {
                                        "description": "url of person",
                                        "default": ""
                                    },
                                    "is_male": {
                                        "description": "gender of person",
                                        "default": ""
                                    },
                                    "birth_date": {
                                        "description": "birth_date of person",
                                        "default": ""
                                    },
                                    "about": {
                                        "description": "about person",
                                        "default": ""
                                    },
                                    "country": {
                                        "description": "birth country of person",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "person created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/people/{id}": {
            "get": {
                "tags": [
                    "person"
                ],
                "summary": "get specified person.",
                "operationId": "42f95ecf49b24da95843e43c429864ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of person",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "person not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "person found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "person"
                ],
                "summary": "update specified person.",
                "operationId": "adbf7ab6502ba622e14ffb6096ed7fe7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of person",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "request body",
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "new name of person",
                                        "default": ""
                                    },
                                    "url": {
                                        "description": "url of person",
                                        "default": ""
                                    },
                                    "birth_date": {
                                        "description": "new birth_date of person",
                                        "default": ""
                                    },
                                    "about": {
                                        "description": "about person",
                                        "default": ""
                                    },
                                    "country": {
                                        "description": "birth country of person",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "person updated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "person not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "person"
                ],
                "summary": "delete specified person.",
                "operationId": "201d5ffb55c341d735d3fc6a8f8dccc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of person",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "person deleted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "person not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/people/{id}/following": {
            "post": {
                "tags": [
                    "person"
                ],
                "summary": "follow specified person.",
                "operationId": "edb2144b1ecdb79bea39292cac37b0ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of person",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "person followed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "person already followed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "person not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "person"
                ],
                "summary": "follow specified person.",
                "operationId": "0081d60529f547fa0b22c40824809433",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of person",
                        "required": true,
                        "schema": {
                            "format": "int64",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "person unfollowed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "person not found or not followed",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}